+++ NEWS.md
@@ -1,3 +1,11 @@
+# MazamaLocationUtils 0.1.10
+
+* Added `table_initializeExisting()` for fast conversion of an existing 
+table of spatial metadata into a standardized "known location" table.
+* Added `addressService` argument to `table_addLocation()`, 
+`table_addSingleLocation() and `location_initialize()` to skip the address
+step that requires web services.
+
 # MazamaLocationUtils 0.1.9
 
 * Documentation tweaks.